tests/wpt/tests/html/semantics/scripting-1/the-script-element/async_009.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Document.write() silently fails from an Async script</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures Document.write() silently fails from an Async script." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/#execute-the-script-block"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script type="text/javascript">
        var t = async_test("Document.write() silently fails from an Async script");

        var log = t.step_func(function() {
          document.write("<span id='writtenText'></span>");
          assert_equals(null, document.getElementById('writtenText'));
          t.done();
        });
        </script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript" src="log.py?sec=1&id=1" async></script>
    </body>
</html>
